requirements.txt
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
Django==6.0.2
django-cors-headers==4.9.0
django-environ==0.12.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
redis==7.1.1
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
google-auth>=2.29.0
gunicorn
whitenoise==6.9.0

